Slightly improve entropy performance
